refactor: extract mirror network constant
